MediaElementRenderer.cs -> Remove If IsPlaying
Hi, I think it would be better to remove the if IsPlaying condition in order to be able to get the position even when the media is paused. Or maybe do something like if IsPaused or IsPlaying. Right now, when the media is paused we get 0 no matter where we are in the media because of the if.